# Use large blocks to optimize building the neighbor list