… platforms/opencl/src/kernels/findInteractingBlocks_cpu.cl