This is an attempt to improve the scaling on large systems.  As discussed in #4063, one part of building the neighbor list involves comparing the bounding boxes of "atom blocks" composed of 32 atoms each.  This scales as O(n^2).  For small to medium systems it takes very little time, but for systems with more than about a million atoms it can become a bottleneck.

This PR introduces "large blocks" composed of 32 atom blocks (1024 atoms).  Each atom block is checked against large blocks to see if we need to check against the individual blocks it contains.  For very large systems, this should reduce the cost of the O(n^2) loop by roughly a factor of 32.  In principle we could add still more levels (blocks of blocks of blocks) to create a full bounding volume hierarchy, which could potentially lead to O(n log n) scaling.

In the current code, atom blocks are sorted by size before building the neighbor list.  Due to the way neighbor lists are represented, this leads to a more efficient list and fewer interactions that have to be computed.  Unfortunately, that's incompatible with this optimization, since large blocks would not be spatially compact.  Therefore, this change is expected to make building the neighbor list faster, but computing interactions slightly slower.

To test this change, I simulated water boxes of varying sizes.  All simulations used reaction field with a cutoff of 1 nm, so that everything other than building the neighbor list should have O(n) scaling.  The graph show the time per 100,000 atoms to simulate 5000 steps.  For perfectly linear scaling this would be constant.

![image](https://github.com/openmm/openmm/assets/4379786/67f6c233-72a3-44e2-8678-93348eec64e1)

The scaling is greatly improved.  It's much faster for the largest systems.  Surprisingly, it also is faster even for the very smallest system tested (only 12,255 atoms).

I also tested it on some of our standard benchmarks, all of which use PME.

|Benchmark|Atoms|Original|New|Ratio|
|---|---|---|---|---|
pme|23558|2036.145|1694.8|0.83
apoa1pme|92224|602.358|627.88|1.04
amber20-cellulose|408609|153.6265|160.629|1.05
amber20-stmv|1067095|44.54215|48.19275|1.08

In this case, the very smallest benchmark became slower, while all others became faster.  Possibly that's because computing interactions is more expensive with PME than with reaction field?

If we automatically switch between the two methods at around 50,000 atoms, that would probably give good results in all cases.

cc @philipturner @bdenhollander @aizvorski