@philipturner 2023-07-18
> In the current code, atom blocks are sorted by size before building the neighbor list.

Would this change remove the GPU -> CPU dependency, so the neighbor list can be entirely GPU-driven? I'm doing something related with a ray tracing BVH, using 16384-atom ($4^3$ nm) blocks. The CPU pre-allocates a heap beforehand, which the GPU creates sub-allocations from. If the GPU runs out of memory, it simply stalls 3 frames to synchronize with the CPU.

<img width="599" alt="Simulation Libraries Maybe" src="https://github.com/openmm/openmm/assets/71743241/1d758a55-9a9b-438f-98fb-01f67b7a7f6f">

This kind of thinking is necessary for using OpenMM in real-time settings, where having the CPU wait on GPU results means UI hitches. For medium-small systems (the ones viable in real-time), CPU<->GPU latency can even dominate the execution time.