@peastman 2023-07-18
It has no effect on GPU-CPU communication.  This change is an optimization for very large systems, not small ones.