@philipturner 2023-07-18
> Unfortunately, that's incompatible with this optimization, since large blocks would not be spatially compact.

This might be possible if you could allocate memory at a finer granularity, GPU-side.

> Therefore, this change is expected to make building the neighbor list faster, but computing interactions slightly slower.

In the second level of my BVH, cells vary from 4/8 - 4/30 nm. That's 25-250 atom references/cell for solid diamond. However, less dense substances like water might be even cheaper.