@peastman 2023-07-18
> This might be possible if you could allocate memory at a finer granularity, GPU-side.

There's no connection.  For this optimization to be useful, each large block (consisting of 32 regular atom blocks) must be spatially compact.  Otherwise they have huge bounding boxes, and aren't useful as a first level filter.  If the blocks are sorted by size instead of position, they aren't compact.