@philipturner 2023-07-18
Could OpenMM incorporate a uniform grid BVH to cheaply sort atoms by position? For example, you could extract ~11 1024-blocks from a 4x4x4 voxel of solid diamond.