@RaulPPelaez 2023-07-18
Since you are playing around with the idea of a full blown LBVH this one is the most performant one I am aware of: https://www.sciencedirect.com/science/article/pii/S092702561930206X?via%3Dihub