@bdenhollander 2023-07-18
This looks promising! If this is ported to OpenCL, I can test on AMD.