@philipturner 2023-07-18
> Since you are playing around with the idea of a full blown LBVH this one is the most performant one I am aware of:

An issue with uniform grids, is they are typically dense. I'm working to make them 2-level sparse. This approach unlocks optimizations like varying granularity of the second level, which helps balance compute cost for non-uniform distributions.

In the paper, the BVH took more time to construct than the grid. MD uses a different approach than RT (DDAs), so it would be helpful to understand why grid traversal is slower.