@peastman 2023-07-18
The atoms are already sorted by position.  That's why this optimization works.  The existing neighbor list building method resorts the atom blocks by size, which would break it.  That's why I have to disable it when using this optimization:

```c++
if (!useLargeBlocks)
    blockSorter->sort(sortedBlocks);
```

> If this is ported to OpenCL, I can test on AMD.

I'm going to look into it.  The CUDA implementation makes use of a `ballot()` function, which isn't available in OpenCL.  The alternative is to use local memory for communicating between the threads, but that's higher overhead.  So this might end up being for CUDA only.  It would also work fine in HIP.