@peastman 2023-07-18
> IIUC there is no code. You expect the user to pre-sort atoms at a 32-atom granularity.

We sort the atoms every 250 steps.  It's done in `ComputeContext::reorderAtoms()`.  Also see the discussion in the [Developer Guide](http://docs.openmm.org/latest/developerguide/06_opencl_platform.html#reordering-of-particles).

> There are OpenCL 2.0 extensions for SIMD ballots.

What are the extensions?