@philipturner 2023-07-18
> What are the extensions?

https://github.com/philipturner/opencl-metal-stdlib#features