@peastman 2023-07-20
I ported it to OpenCL.  Here are the times to simulate 500 steps for the water boxes on M1 Pro.

![image](https://github.com/openmm/openmm/assets/4379786/92061b8c-7bb6-440a-a0bb-3417c8d0d48b)

|Benchmark|Atoms|Original|New|Ratio|
|---|---|---|---|---|
pme|23558|157.54|139.637|0.886359019931446
apoa1pme|92224|38.4132|39.9721|1.0405824039653
amber20-cellulose|408609|7.45802|9.61445|1.28914242654217
amber20-stmv|1067095|1.93384|3.2664|1.68907458734952

It's similar to what we saw for CUDA on the RTX 4080, but the speedups are even larger.