@peastman 2023-07-20
Here are the numbers for Radeon RX 5700 XT.

![image](https://github.com/openmm/openmm/assets/4379786/5c1e4d52-a10b-4cea-9859-d427b6466dab)

|Benchmark|Atoms|Original|New|Ratio|
|---|---|---|---|---|
pme|23558|385.581|248.548|0.644606451044009
apoa1pme|92224|110.795|105.366|0.950999593844487
amber20-cellulose|408609|20.9923|24.7377|1.17841780081268
amber20-stmv|1067095|5.11277|7.4702|1.46108665165849

The general picture is the same for all three GPUs.  The water boxes are faster across the board, with the speedup increasing steadily with system size for large systems.  For the PME benchmarks, DHFR becomes slower on all three GPUs, while Cellulose and STMV are faster on all three.  ApoA1 becomes about 4% faster on the 4080 and M1 Pro, but about 5% slower on the Radeon.  On all three GPUs, the water box that is roughly the same size as ApoA1 has minimal change in speed.  That suggests a cutoff of around 90,000 atoms would make sense for switching between the two behaviors.