@philipturner 2023-07-20
> That suggests a cutoff of around 90,000 atoms would make sense for switching between the two behaviors.

I am not able to implement switching heuristics in my Metal plugin, because the number of atoms isn't known when the `MetalContext` constructor invokes. Code located inside `github.com/openmm/openmm` needs to execute such a heuristic before calling into `philipturner/openmm-metal`, just like with `useNeighborList`. In the meantime, both features are set manually through environment variables.