@philipturner 2023-07-21
Reached a new performance maximum for apoa1rf: 152 -> 157 ns/day. With SIMD ballots and half precision, the overhead is so little, there’s no need for a switching heuristic. It varies between $\pm 4 \\%$ faster/slower for small systems, but <b>isn't consistently worse.</b>

<details>
<summary>Standard Benchmarks</summary>

Benchmark | Atoms | ns/day (Original) | ns/day (1024-blocks)
-- | -- | -- | --
gbsa (No NNL) | 2489 | 858.1 | n/a
gbsa (NNL) | 2489 | 632.0 | 659.5
rf | 23558 | 470.6 | 463.6
pme | 23558 | 267.8 | 262.1
apoa1rf | 92224 | 153.3 | 157.6
apoa1pme | 92224 | 80.1 | 80.7
amber20-cellulose | 408609 | 20.2 | 20.8
amber20-stmv | 1067095 | 7.23 | 7.51



</details>


<details>
<summary>Disabling Neighbor List (Reduces CPU Bottleneck)</summary>

Small Box Sizes (No NNL) | Atoms | ns/day (Original) | ns/day (1024-blocks)  
-- | -- | -- | --
2.0 nm | 774 | 1020 | n/a
2.5 nm | 1503 | 957 | n/a
3.0 nm | 2661 | 820 | n/a
3.5 nm | 4158 | 627 | n/a
4.0 nm | 6282 | 407 | n/a
4.5 nm | 8844 | 257 | n/a
5.0 nm | 12255 | 163 | n/a



</details>

<details>
<summary>Small Box Sizes</summary>

Small Box Sizes (NNL) | Atoms | ns/day (Original) | ns/day (1024-blocks) |
-- | -- | -- | -- 
2.0 nm | 774 | 877 | 892
2.5 nm | 1503 | 795 | 813
3.0 nm | 2661 | 671 | 689
3.5 nm | 4158 | 590 | 605
4.0 nm | 6282 | 490 | 493
4.5 nm | 8844 | 389 | 405
5.0 nm | 12255 | 330 | 337
6.0 nm | 21384 | 297 | 305
7.0 nm | 33840 | 213 | 216

</details>

<details>
<summary>Large Box Sizes</summary>

Large Box Sizes (NNL) | Atoms | s/100K/500 (Original) | s/100K/500 (1024-blocks) |
-- | -- | -- | --
5.0 nm / 5000 steps | 12255 | 4.29 | 4.21
10.0 nm / 5000 steps | 98880 | 2.02 | 2.02
15.0 nm / 2500 steps | 335025 | 1.95 | 1.90
20.0 nm / 1250 steps | 792450 | 2.03 | 1.92
25.0 nm / 600 steps | 1550160 | 2.35 | 2.09
30.0 nm / 300 steps | 2682600 | 2.61 | 2.32

</details>

It is strange that M1 Max gets the exact same speed as M1 Pro, despite having double the compute and bandwidth.