@bdenhollander 2023-07-22
Here are results for Radeon RX 6600 OpenCL on Windows. 

Large blocks was set to true for all tests rather than based on 100000 atoms to confirm the cutoff point. It does appear to be a reasonable value.

| Benchmark          | Atoms   | Original | New     | Ratio  |
|--------------------|---------|----------|---------|--------|
| pme                | 23558   | 304.342  | 235.449 | 0.7736 |
| apoa1pme           | 92224   | 77.6963  | 81.0731 | 1.0435 |
| amber20-cellulose  | 408609  | 18.0175  | 22.7871 | 1.2647 |
| amber20-stmv       | 1067095 | 4.65864  | 6.968   | 1.4957 |

Water boxes

| Width (nm) | Atoms   | Original ns/day | New ns/day | Ratio  |
|------------|---------|----------|------|--------|
| 4          | 6282    | 499      | 506  | 1.0140 |
| 5          | 12255   | 381      | 393  | 1.0315 |
| 6          | 21384   | 341      | 342  | 1.0029 |
| 7          | 33840   | 221      | 217  | 0.9819 |
| 8          | 50442   | 151      | 151  | 1.0000 |
| 9          | 72279   | 109      | 110  | 1.0092 |
| 10         | 98880   | 85.7     | 87.2 | 1.0175 |
| 11         | 131445  | 69.2     | 74   | 1.0694 |
| 15         | 335025  | 24.5     | 29.7 | 1.2122 |
| 20         | 792450  | 7.47     | 11.9 | 1.5930 |
| 25         | 1550160 | 2.62     | 5.46 | 2.0840 |
| 30         | 2682600 | 1.05     | 2.93 | 2.7905 |